Changed main path
The main path changed to 'full', which is more comparable to the former method (Kummer et al., 2020).
A segment minimum length can now be added in milliseconds (set to 20 ms).
Optimized the timing for some variables to ensure they will be stored, even though the data is insufficient.
Added output fields for values relative to the full data epoch (in case of 'nonneg').