Update Statistics_Error_GMA Normalize.R
EEG data is now normalized within the output data frame. Before, it was done in a separate frame which has not been megred with the output matrix. Thus, before, normalization was not applied properly.

Normalization was changed to be done within task, component, gma_measure (as before) and electrode (new), as we analyse electrodes separately later on.