Performance below chance level removed
Participants with an accuracy below chance level within on task will be removed. The accuracy will be calculated from the pertinent behavioral data, only (i.e., absent presenter for Flanker and the speed condition for Go/NoGo).
Added onset_ms and offset_ms to the analyses.
The minimum correlation of the fitted gamma PDF is checked against "equal or greater than" instead of greater than.